requirements.txt
biopython==1.84
botocore==1.34.127
colorama==0.4.6
docutils==0.16
numpy==2.1.1
polars==1.7.1
pyasn1==0.6.0
rsa==4.7.2
setuptools==72.1.0
wheel==0.44.0
